% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dependogram.R
\name{Dependogram}
\alias{Dependogram}
\title{Dependogram for Cramer-von Mises statistics}
\usage{
Dependogram(object, stat, rot = 0)
}
\arguments{
\item{object}{List of the output (statistics, pvalues) from cvm_2series and cvmr_3series}

\item{stat}{Name (string) of statistics to be used}

\item{rot}{Rotation of labels  (default=0)}
}
\value{
\item{Output}{No values are returned; only the graph is printed}
}
\description{
This function, used in cvm_2series and cvm_3series draws the P-values of the Moebius Cramer-von Mises statistics.
}
\references{
Duchesne, Ghoudi & Remillard  (2012). On Testing for independence between the innovations of several time series. CJS, vol. 40, 447-479.
}
